requirements.txt
streamlit>=1.32.0
pyyaml>=6.0
pandas>=2.0
